fix: File sharing feature fixed
