+++ frontend/src/components/SharedContextMenu.jsx
@@ -25,7 +25,7 @@ const SharedContextMenu = ({ file }) => {
         },
         responseType: "blob",
       });
-      console.log(response);
+      
       const blob = await response.data;
       const link = document.createElement("a");
       link.href = URL.createObjectURL(blob);
